Fix types in signedness conversion
Previous code cast to unsigned by default, want signed here.  Addressed a few warnings as well by making casts explicit.

+++ src/player/Player.cpp
@@ -31,11 +31,11 @@ void Player::goTo(unsigned int n) {
 
 /* Set the player's x position within the Hall (relative) */
 void Player::step(int n) {
-    if (posX + n < 0) {
+    if (static_cast<int>(posX) + n < 0) {
         // Program is trying to send to a negative position - send to start instead
         goTo(0);
     } else {
-        goTo(posX + n);
+        goTo(posX + static_cast<unsigned int>(n));
     }
 }
